+++ src/common/resources/index.js
@@ -1,6 +1,7 @@
 export * from './acqUnit';
 export * from './batchGroups';
 export * from './batchVoucherExportsResource';
+export * from './batchVouchers';
 export * from './configAddress';
 export * from './configAdjustments';
 export * from './configApprovals';